requirements.txt
ollama
PyYAML
colorama
python-dotenv
Levenshtein
